requirements.txt
anthropic
PyPDF2
reportlab
Pillow
python-dotenv